Fix an issue with table hashing
When attempting to find a free slot, we were using the first entry with
an empty value. However, if this free slot was a value which had then
then been cleared, then it could be part of the slot "linked list".
Attempting to store in this slot corrupted the linked list, resulting in
the Lua VM crashing.

Instead, we check the key is not nil. We now longer clear the key of a
node when it is garbage collected, to allow us to check whether it is a
free slot or just one which is no longer used.