Inline Math.*Exact methods
While we may lose out on some performance gain from them previously
being intrinsics, the throwing of ArithemticException was causing
serious performance regressions in very specific code paths.

The included code (primes.lua) would cause an overflow (throwing an
exception), and then modulus it (making it wrap back to an integer
again). As a result, this code ran in over 14s (almost instantly
after this patch and on master).