Run all coroutines on a single thread
Coroutines are currently run on separate threads, with yields/resumes
just suspending one thread and allowing another to run again. While this
is fine for small workloads, it has non-trivial resource costs,
especially when users are creating lots of short-lived coroutines.

Instead of running things on separate threads, a context switch between
coroutines (a yield or a resume) is handled by throwing an
UnwindThrowable. This is caught by the main run method, which switches
to another coroutine and resumes there.

Resuming information is based on the pre-existing stack frame. As of the
recent interpreter restructuring we use these frames to re-enter
functions after calling other interpreted functions. It's relatively
trivial to adapt this mechanism for re-entering after a yield too.

"Native" functions are much harder, as we have to manage the state
ourselves, and adapt the existing algorithms to allow re-entry. This can
get a little messy in places, so we limit the areas yielding is allowed.

One extension beyond normal Lua is that we allowed yielding within debug
hooks. This adds substantial complexity, especially for line/instruction
hooks, as we need to a way to ensure our code integrates nicely with the
existing VM.

In order to do this (and several other, similar problems), we make
liberal use of PUC Lua's "call info state" flags. This marks several
boolean properties about the current frame, such as whether it's in a
hook, whether it errored, etc...

As part of this rewrite, we've also fixed several issues with debug
hooks, which means we can uncomment several more test cases!