Use ropes for string concatenation
When concatinating two strings (a .. b), we build an intermediate rope
structure. This is then realised into an actual string when attempting
to perform other operations on it (string.* functions, mathematical
operations, etc...).

As rope concatination is O(1)*, this ends up being faster when doing
multiple concatinations in a loop (without operating on the string in
other ways). For instance:

    local a = ""
    for i = 1, n do a = a .. i .. ", " end

This loop now operates in O(n) rather than O(n^2). Note that this change
shouldn't have any major impact on normal concatination (a .. b .. c ..
etc...) as that's already optimised. That said, I've done little
profiling on those cases, so don't know yet.

* Strictly speaking its O(lg n), as we should do rebalancing. However,
  that's not been implemented yet.