Use ropes for string concatenation (#44)
When concatenating two strings (a .. b), we build an intermediate rope
structure. This is then realised into an actual string when attempting
to perform other operations on it (string.* functions, mathematical
operations, etc...).

As rope concatenation is O(1)*, this ends up being faster when doing
multiple concatenations in a loop (without operating on the string in
other ways). For instance:

    local a = ""
    for i = 1, n do a = a .. i .. ", " end

This loop now operates in O(n) rather than O(n^2). It's still often
more efficient to use table.concat, but now by a constant(ish) factor.