WIP Doubles: Initial work on porting google's double-conversion
The code is still mostly C++, but there's a method to the madness!
Keeping most of the original code will assist in verifying the
correctness of the port.

The header files were combined with compilation units, to make java
classes.  Positions were kept as close to the compilation unit as
possible to facilitate direct diff comparisons.