Made most things work again
I had to revert the changes to the code generator back to pure 5.1, so
_ENV is no longer special in code. However, the environment is still
stored as an upvalue, even on 5.1 (which is the version the bytecode
is generated for). It properly loads 5.2 bytecode for the most part, but
there's still a few things broken with upvalues. (It's able to get CC
to partially boot through the BIOS!)