doubles: wrap BigInteger instead of BigDecimal to greatly improve performance
I misunderstood exactly what the original BigNum was doing, and assumed
it was another floating point container.  It turns out it uses the
exponent to improve perfomance on certain calculations.

I switched to wrapping a BigInteger(which seems to have smilar
optimizations), and improved performance by orders of magnitude.

Around 18,000 calculations went from ~1 minute to ~4 seconds. (in DtpaTest.dtoaGayPrecision)