Proper coroutines, properly
A few years ago (2019) we rewrote Cobalt's coroutine implementation to
use "proper" coroutines (i.e. unwinding the stack on a yield), rather
than using a Java thread for each coroutine.

However, behind the scenes, coroutine.yield would still be implemented
by switching threads in the event that a non-unwindable Java function
was on the stack. Examples of this include CC's Lua functions and
~load~.

As of MC 1.13, CC's Lua functions handle unwinding properly, which means
that ~load~ is the only remaining non-unwindable function. There's a
reason for this: if the reader inside load yields, this means we need to
suspend the whole Lua compiler, a pretty daunting task!

It turns out it's actually incredibly easy to do this, the trick is just
not to do it by hand. There's a lot of libraries out there which are
capable of rewriting the bytecode to allow unwinding/rewinding the stack
at arbitrary points. As of this commit, there's one more.

 - Add a new interface for a suspendable function, on top of the
   existing Resumable interface. Unlike Resumable, SuspendedTask and
   SuspendedFunction hold their own state (rather than letting it live
   on the debug frame).

 - Refactor pcall into a separate class, which now calls an arbitrary
   SuspendableFunction in a protected context. This is still a pretty
   nasty interface - it doesn't compose very well - but it's an
   improvement!

 - Create a new build-tools library, which instruments our compiled
   classes, converting classes/functions annotated with an @AutoUnwind
   annotation to support unwinding. The translation scheme is fairly
   simple:

    - Every annotated function f(...) gets transformed to f(..., state :
      Object).

    - Every call to a yielding function gets wrapped in a try {...}
      catch{...}, where the catch block writes the current state, and
      any locals to the captured exception.

    - When resuming a function we pass this state object as the last
      argument to the function. When the state is not-null, we set up
      our locals/stack again and then jump to the function call which
      needs resuming.

   Note that unlike Resumable and the Lua stack, where we only resume on
   the leaf function, resuming Java code sets up the whole Java call
   stack again. This does actually complicate the translation scheme a
   bit, but also I think is the Correct(TM) thing to do.

   There are some pretty big limitations with our translation scheme
   right now - there's limited support for restoring the current stack,
   so the results of yielded functions sometimes need to be stored to
   intermediate locals. It's a bit of a pain, but not a massive hoop to
   jump through.

 - Convert the compiler to use this @AutoUnwind mechanic.

 - Strip out all the threading code from LuaThread.

There's still a lot of work to be done here - I'd like to use this
system for some of the table methods too, but we're a way off that yet.