Merge the rope implementation into LuaString
I'm not 100% convinced this is the right thing to do (especially when it
comes to allocation tracking in the future), but any alternative I can
think of introduces more allocations/pointer indirections, which we
obviously want to avoid!

Also further simplification to the LuaValue interface. Will eventually
tear this out, but let's try to get rid of the worst bits now.