Roll another test framework
I sometimes feel like a parody of myself.

This adds a new test framework to Cobalt, based on CC:T's mcfly and
the existing Busted testing framework. This provides a couple of
advantages over our existing test frameworks:

 - Encourages writing smaller tests, rather than chucking dozens of
   assertions into a single file.

 - Provides a (somewhat) structured way of writing assertions, rather
   than our various ad-hoc assert_eq(...) methods.

 - Runs on native Lua, meaning we can test compatibility with PUC Lua
   using the following Fish shell script:

     for lua in /usr/bin/lua5.{1,2,3,4};
       $lua src/test/resources/spec/_run.lua
     end

Now, this clearly isn't a good unit-test framework - everything is
heavily reliant on the Lua runtime - but that was already the case
before, so I don't think we're losing anything.