Handle timeouts via interrupts rather than polling
Currently the recommended way of handling timeouts is by implementing a
custom DebugHandler and overriding onInstruction/poll. This was probably
a bad idea for a couple of reasons:

 - Every instruction must now call DebugHandler.onInstruction, which
   adds some performance overhead. Ideally the JIT would be able to
   eliminate the virtual call, but we can't rely on that.

 - Suspending the VM inside onInstruction ends up being quite clunky, as
   you need to juggle various debug frame flags.

This commit tears out all of this functionality, and replaces it with an
interrupt-based approach:

 - LuaState.interrupt(): Sets the interrupt flag on the Lua state. This
   is typically called from another watchdog-style thread (such as
   CC:T's ComputerThread.monitor).

 - LuaState.isInterrupted()/LuaState.handleInterrupt(): These are called
   from within the Lua VM at opportune times (currently on instruction,
   though we may change that).

 - InterruptHandler: An interrupt handler is provided to the LuaState
   when constructing it. When handleInterrupt is called, the interrupt
   handler is invoked. The handler may either error, or indicate that it
   wishes to suspend the currently running VM.

This should have a much lower overhead (the hotpath is just a single
load and branch), though I confess I've not actually run any benchmarks
yet. It's definitely a much simpler interface to consume.

This will require some changes to CC:T - while the timeout system is
push-based, there isn't any way to listen to state changes.

This commit also contains a couple of related and unrelated changes, as
I have terrible commit discipline:

 - Add @Serial annotations everywhere

 - Expose LuaThread.status as an enum rather than a string.

 - Clean up the debug hooks a little bit, to be notionally faster on the
   hotpath (when no hooks are active).

   Notionally because I implemented something much simpler a few weeks
   ago and it had very little impact. I'm not sure if something else is
   the dominating factor or if the JIT is being very clever.