Move the system libraries to the test package
The code in these modules is pretty bad compared with the rest of Cobalt
(and that's saying something!). Given these files only exist for the
purpose of tests, let's move them to the test folder.

I've some major refactoring I want to do in the future, and it doesn't
feel worth applying some of those improvements to these files - they
only need to do the bare minimum.