Don't error if we already have a lookahead token
When parsing "if true do goto() end", we'll try to parse the "goto()" as
a goto statement (which requires one lookahead), then backtrack and
parse the whole if body as a full statement (which then tries to parse
as a goto statement again).

As the lookahead token is already present, this causes a crash. To fix
this, we remove this assertion - it makes sense in PUC Lua, but not for
Cobalt's more complex parser.

Fixes #84