Run call/return hooks when calling
Previously we handled pushing/popping frames (and all the other
machinery like debug hooks) inside the function's implementation. This
had several side effects:

 - Most native functions didn't push a frame at all. This was a bit of a
   pain for some of the debug functions, as we had to do funky level
   adjustments, and generally resulted in inconsistencies with PUC Lua.

 - ResumableVarArgFunction and the Lua interpreter had a bunch of
   duplicated logic to handle calling debug hooks.

This instead moves the logic to when we actually call a function, which
is now done in a centralised location (a new Dispatch class)[^1].

There's still some ugliness here, partly due to how we need to set up
frames for the Lua interpreter. This wouldn't be too bad, except we have
5 different forms of call/invoke.

One nice side effect of this is we now implement the "extra error stack"
functionality properly, which means you won't get stack overflows in the
xpcall error handler when recovering from a stack overflow.

There is a definite performance regression with this patch due to us
pushing frames on each library call, but benchmarks show it's not too
terrible except in very tight loops. We could possibly avoid pushing
stack frames when debug hooks are not enabled, but this feels even more
awkward to get right.

[^1] With the exception of the interpreter. Ideally we'd have some
equivalent of preCall(f, args), which returns an Either<DebugFrame,
Varargs>, but without stack allocation I'm not comfortable doing that.